requirements.txt
PyQt6~=6.5.3
Pillow~=10.0.0
requests~=2.31.0
numpy~=1.26.1